KaradagTalaz/requirements.txt
numpy
torch
torchvision
jpegio
scipy
torch_dct
pillow
matplotlib

